+++ Test/SmallCheck.hs
@@ -22,7 +22,6 @@ module Test.SmallCheck (
 
   Testable,
   Property,
-  property,
 
   -- ** Existential quantification